test: simplify TIDA-00553 repro
