--- site/examples/repro-tida00553-j4-cell-divider.page.tsx
@@ -1,4 +1,0 @@
-import { PipelineDebugger } from "site/components/PipelineDebugger"
-import inputProblem from "../../tests/repros/repro-tida00553-j4-cell-divider.input"
-
-export default () => <PipelineDebugger inputProblem={inputProblem} />
